make a ProfileContainer
