+++ src/containers/Profile/ProfileContainer.js
@@ -1,0 +1,13 @@
+import { connect } from "react-redux";
+import getComments from "../../actions/comment/comment-list-action";
+import Profile from "../../components/Profile/Profile";
+
+const mapStateToProps = state => {
+  return state.user;
+};
+
+const mapDispatchToProps = dispatch => ({
+  getComments: (data, page) => dispatch(getComments(data, page)),
+});
+
+export default connect(mapStateToProps, mapDispatchToProps)(Profile);
